remove default pane
