+++ packages/nexusgraph-graph/src/inspection-panel/properties-panel-content/properties-panel-content/DetailsPane.tsx
@@ -2,12 +2,20 @@
 import { useState } from "react";
 import { StyleableRelType } from "./StyleableRelType";
 import { StyleableNodeLabel } from "./StyleableNodeLabel";
-import { DetailsPaneProps, upperFirst } from "../../DefaultDetailsPane";
 import { PaneBody, PaneHeader, PaneTitle, PaneWrapper } from "../../../styles/DefaultPane.styled";
 import { ClipboardCopier } from "../../ClipboardCopier";
 import { PropertiesTable } from "../../PropertiesTable";
+import { NodeItem, RelationshipItem } from "../../../VizItem";
+import { GraphStyleModel } from "../../../GraphStyle";
 
 export const DETAILS_PANE_STEP_SIZE = 1000;
+
+export interface DetailsPaneProps {
+  vizItem: NodeItem | RelationshipItem;
+  graphStyle: GraphStyleModel;
+  nodeInspectorWidth: number;
+}
+
 export function DetailsPane({ vizItem, graphStyle, nodeInspectorWidth }: DetailsPaneProps): JSX.Element {
   const [maxPropertiesCount, setMaxPropertiesCount] = useState(DETAILS_PANE_STEP_SIZE);
